Second iteration of the container: adding isolation via namespaces.
By adding a clone of a new UTS namespace, we can specifically isolate the hostname. Now when we change the hostname while the container is running and then exit the container, the host machine itself still has the original hostname and was not affected by the changes made inside the running container.